Added attrib functionality to the dna_align_feature.
It is now possible to assign attributes to a dna_align_feature, fetch it
and it will be deleted along with the feature if the feature is removed.

The only problem, which is most likely not a problem, it that attribs are not
added to a feture when it is fetched. This have to be done by the user. The only
time this could be a problem, is if a feature is removed and then stored
in the database. If the features has not been fetched at the same time they
will be lost.